Refactor nav.web.auth and nav.django.auth
Use `git blame -C` to see who originally created a line.
