+++ python/nav/web/auth/middleware.py
@@ -17,15 +17,13 @@
 import os
 from logging import getLogger
 
-from nav.config import NAV_CONFIG
 from nav.models.profiles import Account
-from nav.django.utils import is_admin, get_account
-from nav.web.auth import authenticate_remote_user
+from nav.web.auth import authenticate_remote_user, ACCOUNT_ID_VAR
+from nav.web.auth.sudo import get_sudoer, SUDOER_ID_VAR
 
 from django.http import HttpResponseRedirect, HttpResponse
 from django.contrib.sessions.backends.db import SessionStore
 from django.conf import settings
-from django.utils.encoding import python_2_unicode_compatible
 from django.utils.six.moves.urllib import parse
 try:
     from django.utils.deprecation import MiddlewareMixin
@@ -34,9 +32,6 @@ except ImportError:  # Django <= 1.9
 
 
 _logger = getLogger(__name__)
-
-ACCOUNT_ID_VAR = 'account_id'
-SUDOER_ID_VAR = 'sudoer'
 
 # This may seem like redundant information, but it seems django's reverse
 # will hang under some usages of these middleware classes - so until we figure
@@ -53,7 +48,7 @@ class AuthenticationMiddleware(MiddlewareMixin):
             # Get or create an account from the REMOTE_USER http header
             account = authenticate_remote_user(request)
             if account:
-                session[ACCOUNT_ID_VAR]  = account.id
+                session[ACCOUNT_ID_VAR] = account.id
             else:
                 # Set account id to anonymous user
                 session[ACCOUNT_ID_VAR] = Account.DEFAULT_ACCOUNT
@@ -110,83 +105,3 @@ def authorization_not_required(fullpath):
     for url in auth_not_required:
         if fullpath.startswith(url):
             return True
-
-#
-# sudo-related functionality
-#
-
-
-def sudo(request, other_user):
-    """Switches the current session to become other_user"""
-    if SUDOER_ID_VAR in request.session:
-        # Already logged in as another user.
-        raise SudoRecursionError()
-    if not is_admin(get_account(request)):
-        # Check if sudoer is acctually admin
-        raise SudoNotAdminError()
-    request.session[SUDOER_ID_VAR] = request.account.id
-    request.session[ACCOUNT_ID_VAR] = other_user.id
-    request.session.save()
-    request.account = other_user
-
-
-def desudo(request):
-    """Switches the current session to become the original user from before a
-    call to sudo().
-
-    """
-    if not SUDOER_ID_VAR in request.session:
-        # We are not sudoing
-        return
-
-    original_user_id = request.session[SUDOER_ID_VAR]
-    original_user = Account.objects.get(id=original_user_id)
-
-    del request.session[ACCOUNT_ID_VAR]
-    del request.session[SUDOER_ID_VAR]
-    request.session[ACCOUNT_ID_VAR] = original_user_id
-    request.session.save()
-    request.account = original_user
-
-
-def get_sudoer(request):
-    """Returns a sudoer's Account, if current session is in sudo-mode"""
-    if SUDOER_ID_VAR in request.session:
-        return Account.objects.get(id=request.session[SUDOER_ID_VAR])
-
-
-@python_2_unicode_compatible
-class SudoRecursionError(Exception):
-    msg = u"Already posing as another user"
-
-    def __str__(self):
-        return self.msg
-
-
-@python_2_unicode_compatible
-class SudoNotAdminError(Exception):
-    msg = u"Not admin"
-
-    def __str__(self):
-        return self.msg
-
-
-def create_session_cookie(username):
-    """Creates an active session for username and returns the resulting
-    session cookie.
-
-    This is useful to fake login sessions during testing.
-
-    """
-    user = Account.objects.get(login=username)
-    session = SessionStore()
-    session[ACCOUNT_ID_VAR] = user.id
-    session.save()
-
-    cookie = {
-        'name': settings.SESSION_COOKIE_NAME,
-        'value': session.session_key,
-        'secure': False,
-        'path': '/',
-    }
-    return cookie
