+++ python/nav/web/auth/sudo.py
@@ -1,0 +1,91 @@
+#
+# Copyright (C) 2013 Uninett AS
+#
+# This file is part of Network Administration Visualized (NAV).
+#
+# NAV is free software: you can redistribute it and/or modify it under
+# the terms of the GNU General Public License version 3 as published by
+# the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+# details.  You should have received a copy of the GNU General Public License
+# along with NAV. If not, see <http://www.gnu.org/licenses/>.
+#
+"""NAV authentication and authorization middleware for Django"""
+from logging import getLogger
+
+from nav.models.profiles import Account
+from nav.django.utils import is_admin, get_account
+from nav.web.auth import ACCOUNT_ID_VAR
+
+from django.utils.encoding import python_2_unicode_compatible
+
+
+_logger = getLogger(__name__)
+
+SUDOER_ID_VAR = 'sudoer'
+
+# This may seem like redundant information, but it seems django's reverse
+# will hang under some usages of these middleware classes - so until we figure
+# out what's going on, we'll hardcode this here.
+LOGIN_URL = '/index/login/'
+#
+# sudo-related functionality
+#
+
+
+def sudo(request, other_user):
+    """Switches the current session to become other_user"""
+    if SUDOER_ID_VAR in request.session:
+        # Already logged in as another user.
+        raise SudoRecursionError()
+    if not is_admin(get_account(request)):
+        # Check if sudoer is acctually admin
+        raise SudoNotAdminError()
+    request.session[SUDOER_ID_VAR] = request.account.id
+    request.session[ACCOUNT_ID_VAR] = other_user.id
+    request.session.save()
+    request.account = other_user
+
+
+def desudo(request):
+    """Switches the current session to become the original user from before a
+    call to sudo().
+
+    """
+    if SUDOER_ID_VAR not in request.session:
+        # We are not sudoing
+        return
+
+    original_user_id = request.session[SUDOER_ID_VAR]
+    original_user = Account.objects.get(id=original_user_id)
+
+    del request.session[ACCOUNT_ID_VAR]
+    del request.session[SUDOER_ID_VAR]
+    request.session[ACCOUNT_ID_VAR] = original_user_id
+    request.session.save()
+    request.account = original_user
+
+
+def get_sudoer(request):
+    """Returns a sudoer's Account, if current session is in sudo-mode"""
+    if SUDOER_ID_VAR in request.session:
+        return Account.objects.get(id=request.session[SUDOER_ID_VAR])
+
+
+@python_2_unicode_compatible
+class SudoRecursionError(Exception):
+    msg = u"Already posing as another user"
+
+    def __str__(self):
+        return self.msg
+
+
+@python_2_unicode_compatible
+class SudoNotAdminError(Exception):
+    msg = u"Not admin"
+
+    def __str__(self):
+        return self.msg
