+++ python/nav/web/modpython.py
@@ -39,7 +39,8 @@ from django.utils.six.moves import http_cookies
 from nav.bootstrap import bootstrap_django
 
 from django.contrib.sessions.middleware import SessionMiddleware
-from nav.django.auth import AuthenticationMiddleware, AuthorizationMiddleware
+from nav.web.auth.middleware import AuthenticationMiddleware
+from nav.web.auth.middleware import AuthorizationMiddleware
 from nav.web import loginit
 from django.db import connection
