+++ python/nav/web/navlets/__init__.py
@@ -58,8 +58,8 @@ from django.views.generic.base import TemplateView
 
 from nav.models.profiles import AccountNavlet, AccountDashboard
 from nav.models.manage import Sensor
-from nav.django.auth import get_sudoer
 from nav.django.utils import get_account
+from nav.web.auth.sudo import get_sudoer
 from nav.web.utils import require_param
 from nav.web.webfront import find_dashboard
