+++ tests/unittests/general/web_middleware_test.py
@@ -1,13 +1,10 @@
 from mock import patch
 import os
 
-import pytest
-
-from nav.django.auth import ACCOUNT_ID_VAR, SUDOER_ID_VAR
-from nav.django.auth import AuthenticationMiddleware
-from nav.django.auth import AuthorizationMiddleware
-from nav.django.auth import sudo, desudo
-from nav.django.auth import SudoRecursionError, SudoNotAdminError
+from nav.web.auth import ACCOUNT_ID_VAR
+from nav.web.auth.sudo import SUDOER_ID_VAR
+from nav.web.auth.middleware import AuthenticationMiddleware
+from nav.web.auth.middleware import AuthorizationMiddleware
 from nav.web import auth
 
 
@@ -37,7 +34,7 @@ class TestAuthenticationMiddleware(object):
 
     def test_process_request_logged_in(self):
         fake_request = FakeRequest(session={ACCOUNT_ID_VAR: 101})
-        with patch('nav.django.auth.Account.objects.get', return_value=PLAIN_ACCOUNT):
+        with patch('nav.web.auth.Account.objects.get', return_value=PLAIN_ACCOUNT):
             AuthenticationMiddleware().process_request(fake_request)
             assert fake_request.account == PLAIN_ACCOUNT
             assert fake_request.session[ACCOUNT_ID_VAR] == fake_request.account.id
@@ -49,14 +46,14 @@ class TestAuthenticationMiddleware(object):
                 SUDOER_ID_VAR: True,
             }
         )
-        with patch('nav.django.auth.Account.objects.get'):
-            with patch('nav.django.auth.get_sudoer', return_value='foo'):
+        with patch('nav.web.auth.middleware.Account.objects.get'):
+            with patch('nav.web.auth.middleware.get_sudoer', return_value='foo'):
                 AuthenticationMiddleware().process_request(fake_request)
                 assert getattr(fake_request.account, 'sudo_operator', None) == 'foo'
 
     def test_process_request_not_logged_in(self):
         fake_request = FakeRequest(session={})
-        with patch('nav.web.auth.Account.objects.get', return_value=DEFAULT_ACCOUNT):
+        with patch('nav.web.auth.middleware.Account.objects.get', return_value=DEFAULT_ACCOUNT):
             AuthenticationMiddleware().process_request(fake_request)
             assert fake_request.account == DEFAULT_ACCOUNT
             assert fake_request.session[ACCOUNT_ID_VAR] == fake_request.account.id
@@ -65,7 +62,7 @@ class TestAuthenticationMiddleware(object):
         fake_request = FakeRequest(
             META={'REMOTE_USER': 'tim'}
         )
-        with patch('nav.django.auth.authenticate_remote_user', return_value=PLAIN_ACCOUNT):
+        with patch('nav.web.auth.middleware.authenticate_remote_user', return_value=PLAIN_ACCOUNT):
             AuthenticationMiddleware().process_request(fake_request)
             assert fake_request.account == PLAIN_ACCOUNT
             assert fake_request.session[ACCOUNT_ID_VAR] == fake_request.account.id
@@ -80,67 +77,23 @@ class TestAuthorizationMiddleware(object):
     def test_process_request_anonymous(self):
         fake_request = FakeRequest()
         fake_request.account = DEFAULT_ACCOUNT
-        with patch('nav.django.auth.authorization_not_required', return_value=True):
+        with patch('nav.web.auth.middleware.authorization_not_required', return_value=True):
             AuthorizationMiddleware().process_request(fake_request)
             assert 'REMOTE_USER' not in os.environ
 
     def test_process_request_authorized(self):
         fake_request = FakeRequest()
         fake_request.account = PLAIN_ACCOUNT
-        with patch('nav.django.auth.authorization_not_required', return_value=True):
+        with patch('nav.web.auth.middleware.authorization_not_required', return_value=True):
             AuthorizationMiddleware().process_request(fake_request)
             assert os.environ.get('REMOTE_USER', None) == PLAIN_ACCOUNT.login
 
     def test_process_request_not_authorized(self):
         fake_request = FakeRequest()
         fake_request.account = PLAIN_ACCOUNT
-        with patch('nav.django.auth.authorization_not_required', return_value=False):
-            with patch('nav.django.auth.Account.has_perm', return_value=False):
-                with patch('nav.django.auth.AuthorizationMiddleware.redirect_to_login', return_value='here'):
+        with patch('nav.web.auth.middleware.authorization_not_required', return_value=False):
+            with patch('nav.web.auth.Account.has_perm', return_value=False):
+                with patch('nav.web.auth.middleware.AuthorizationMiddleware.redirect_to_login', return_value='here'):
                     result = AuthorizationMiddleware().process_request(fake_request)
                     assert result == 'here'
                     assert os.environ.get('REMOTE_USER', None) != PLAIN_ACCOUNT.login
-
-
-class TestSudo(object):
-
-    def test_sudo_already(self):
-        fake_request = FakeRequest(session={SUDOER_ID_VAR: True})
-        with pytest.raises(SudoRecursionError):
-            sudo(fake_request, None)
-
-    def test_sudo_not_admin(self):
-        fake_request = FakeRequest()
-        with pytest.raises(SudoNotAdminError):
-            with patch('nav.django.auth.is_admin', return_value=False):
-                with patch('nav.django.auth.get_account', return_value=None):
-                    sudo(fake_request, None)
-
-    def test_sudo_ok(self):
-        fake_request = FakeRequest(
-            session={ACCOUNT_ID_VAR: PLAIN_ACCOUNT.id}
-        )
-        fake_request.account = PLAIN_ACCOUNT
-        with patch('nav.django.auth.is_admin', return_value=True):
-            with patch('nav.django.auth.get_account', return_value=None):
-                sudo(fake_request, DEFAULT_ACCOUNT)
-                assert fake_request.account == DEFAULT_ACCOUNT
-                assert fake_request.session[ACCOUNT_ID_VAR] == DEFAULT_ACCOUNT.id
-                assert fake_request.session[SUDOER_ID_VAR] == PLAIN_ACCOUNT.id
-
-
-class TestDesudo(object):
-
-    def test_desudo(self):
-        fake_request = FakeRequest(
-            session={
-                SUDOER_ID_VAR: PLAIN_ACCOUNT.id,
-                ACCOUNT_ID_VAR: DEFAULT_ACCOUNT.id,
-            }
-        )
-        fake_request.account = DEFAULT_ACCOUNT
-        with patch('nav.django.auth.Account.objects.get', return_value=PLAIN_ACCOUNT):
-            desudo(fake_request)
-            assert fake_request.account == PLAIN_ACCOUNT
-            assert fake_request.session[ACCOUNT_ID_VAR] == PLAIN_ACCOUNT.id
-            assert SUDOER_ID_VAR not in fake_request.session
